memory/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
upstash-redis==0.15.0
psycopg2-binary==2.9.9
pymongo[srv]>=4.7.0
python-dotenv==1.0.0
httpx==0.25.2
